revise BinanceTx type to match type of signed data
The data signed on the Binance chain is a JSON structure, and BinanceTx
should match its type. The important differences are that the data field
can be (in fact, is right now always) null, and that the amounts are
passed in as numbers rather than strings.

This commit also adds BNB testnet support for the native wallet, and
ensures that the txid and serialized fields are returned.